requirements.txt
biopython
fastai
fastparquet
gffutils
ipywidgets
jupyterlab
jupyterlab-quarto
lxml
nbdev
requests
seaborn
torchtext
tqdm